reduce mem in caption embedder

+++ examples/opensora_hpcai/opensora/models/layers/blocks.py
@@ -653,7 +653,7 @@ class CaptionEmbedder(nn.Cell):
 
         # manually expand dims to avoid infer-shape bug in ms2.3 daily
         caption = ops.where(
-            drop_ids[:, None, None, None], self.y_embedding[None, None, :, :], caption.to(self.y_embedding.dtype)
+            drop_ids[:, None, None, None], self.y_embedding[None, None, :, :].to(caption.dtype), caption)
         )
 
         return caption
